Refact/processes (#94)
* :truck: moved spool initialization in server creation.

* :recycle: moved processes into server initialization.
	  moved spool folder creation into server initialization.
	  created a Queue file for the queue struct.

* :arrow_up: handled send errors in queue's message handling.
	   updated tests.
